refactor: move solution, osrmhelper and searchhelper into oasis/internal issue: https://github.com/Telenav/osrm-backend/issues/342
