+++ integration/service/oasis/selectionstrategy/reachable_by_single_charge.go
@@ -10,8 +10,8 @@ import (
 	"github.com/Telenav/osrm-backend/integration/api/osrm"
 	"github.com/Telenav/osrm-backend/integration/api/osrm/table"
 	"github.com/Telenav/osrm-backend/integration/api/search/nearbychargestation"
+	"github.com/Telenav/osrm-backend/integration/service/oasis/internal/osrmhelper"
 	"github.com/Telenav/osrm-backend/integration/service/oasis/osrmconnector"
-	"github.com/Telenav/osrm-backend/integration/service/oasis/osrmhelper"
 	"github.com/Telenav/osrm-backend/integration/service/oasis/stationfinder/stationfinderalg"
 	"github.com/golang/glog"
 )
